Implements PlayerDao and CityDao, change int to BigInteger, create tests for PLayerDao and CityDao

+++ src/main/java/com/nctc2017/dao/CityDao.java
@@ -1,12 +1,13 @@
 package com.nctc2017.dao;
 
+import java.math.BigInteger;
 import java.util.*;
 
 import com.nctc2017.bean.City;
 
 public interface CityDao {
 
-    City find(int cityId);
+    City find(BigInteger cityId);
 
     List<City> findAll();
